refactor utils to export helper functions
