+++ ui/src/app/panel/thread-list/thread-list.component.ts
@@ -5,7 +5,7 @@ import {Observable} from 'rxjs';
 import {switchMap} from 'rxjs/operators';
 
 import {DataService} from 'app/service/data.service';
-import {UtilsService} from 'app/service/utils.service';
+import {convertTimestamp} from 'app/service/utils';
 
 @Component({
   templateUrl: './thread-list.component.html',
@@ -13,17 +13,15 @@ import {UtilsService} from 'app/service/utils.service';
 })
 export class ThreadListComponent implements OnInit {
   threadList$: Observable<any>;
-  selectedId: number;
+  convertTimestamp = convertTimestamp;
 
   constructor(private dataService: DataService,
-              private route: ActivatedRoute,
-              private utils: UtilsService) {
+              private route: ActivatedRoute) {
   }
 
   ngOnInit(): void {
     this.threadList$ = this.route.paramMap.pipe(
       switchMap(params => {
-        this.selectedId = +params.get('id');
         return this.dataService.threadList;
       })
     );
