+++ ui/src/app/search-bar/search-bar.component.ts
@@ -2,6 +2,7 @@ import {COMMA, ENTER} from '@angular/cdk/keycodes';
 import {Component, OnInit} from '@angular/core';
 import {MatChipInputEvent} from '@angular/material/chips';
 import {Router} from '@angular/router';
+
 import {DataService} from 'app/service/data.service';
 
 @Component({
